fix: address PR #70 review issues
- Remove `as any` casts: use typed `GatewayToolEndEvent['images']` in
  agent-bridge.ts, add `SendableChannel` type guard in discord.ts
- Extract shared CSS→DIP capture utility to electron/utils/capture.ts,
  eliminating duplication between ipc/app-control.ts and cli/app-control.ts
- Extract agent/chat/model types from ipc.ts (597→475 LOC) into
  src/types/agent.ts, re-exported for backward compat
- Extract provider metadata maps to electron/ipc/provider-metadata.ts,
  restoring one-entry-per-line formatting for clean diffs
- Fix ImageLightbox keyboard shortcuts not working on open: add ref
  callback to auto-focus overlay on mount
- Remove unused `execFileSync` import from rebuild-better-sqlite3.mjs
