+++ apps/desktop/electron/ipc/agent-helpers.ts
@@ -324,33 +324,8 @@ export function attachmentsToImages(attachments?: ChatAttachment[]): ImageConten
 
 // ── Provider metadata ────────────────────────────────────────
 
-/* eslint-disable @typescript-eslint/naming-convention -- provider ID keys */
-const PROVIDER_NAMES: Record<string, string> = {
-  anthropic: 'Anthropic', openai: 'OpenAI', 'openai-codex': 'OpenAI (Codex)',
-  google: 'Google', 'google-gemini-cli': 'Google (Gemini CLI)',
-  'google-antigravity': 'Antigravity', 'google-vertex': 'Google Vertex',
-  xai: 'xAI', openrouter: 'OpenRouter', groq: 'Groq', cerebras: 'Cerebras',
-  mistral: 'Mistral', 'github-copilot': 'GitHub Copilot',
-  'amazon-bedrock': 'Amazon Bedrock', 'azure-openai-responses': 'Azure OpenAI',
-  huggingface: 'Hugging Face', 'vercel-ai-gateway': 'Vercel AI Gateway',
-  zai: 'ZAI', opencode: 'OpenCode', 'kimi-coding': 'Kimi', minimax: 'MiniMax',
-  'minimax-cn': 'MiniMax CN',
-};
-const PROVIDER_LOGO_MAP: Record<string, string> = {
-  'openai-codex': 'openai', 'google-gemini-cli': 'google',
-  'google-antigravity': 'google', 'google-vertex': 'google-vertex',
-  'azure-openai-responses': 'azure', 'amazon-bedrock': 'amazon-bedrock',
-  'github-copilot': 'github-copilot', 'vercel-ai-gateway': 'vercel',
-  'kimi-coding': 'openai',
-};
-/* eslint-enable @typescript-eslint/naming-convention */
-
-export function providerLogo(provider: string): string {
-  return `https://models.dev/logos/${PROVIDER_LOGO_MAP[provider] ?? provider}.svg`;
-}
-export function providerDisplayName(provider: string): string {
-  return PROVIDER_NAMES[provider] ?? provider.replace(/-/g, ' ').replace(/\b\w/g, (c) => c.toUpperCase());
-}
+import { providerLogo, providerDisplayName } from './provider-metadata';
+export { providerLogo, providerDisplayName };
 
 // ── Model state builder ──────────────────────────────────────
